use f32 as standard and allow switching to f64 using a feature

+++ examples/simple-categorical.rs
@@ -3,6 +3,7 @@ extern crate esopt;
 
 use ofnn::*;
 use esopt::*;
+use ofnn::Float;
 
 
 fn main()
@@ -60,13 +61,13 @@ fn main()
 struct NNEvaluator
 {
     model:Sequential,
-    x:Vec<Vec<f64>>,
-    y:Vec<Vec<f64>>,
+    x:Vec<Vec<Float>>,
+    y:Vec<Vec<Float>>,
 }
 
 impl NNEvaluator
 {
-    pub fn new(model:Sequential, x:Vec<Vec<f64>>, y:Vec<Vec<f64>>) -> NNEvaluator
+    pub fn new(model:Sequential, x:Vec<Vec<Float>>, y:Vec<Vec<Float>>) -> NNEvaluator
     {
         NNEvaluator { model: model, x: x, y: y }
     }
@@ -74,7 +75,7 @@ impl NNEvaluator
 
 impl Evaluator for NNEvaluator
 {
-    fn eval_test(&self, params:&[f64]) -> f64
+    fn eval_test(&self, params:&[Float]) -> Float
     {
         let mut local = self.model.clone();
         local.set_params(params);
@@ -87,7 +88,7 @@ impl Evaluator for NNEvaluator
         score
     }
     
-    fn eval_train(&self, params:&[f64], _:usize) -> f64
+    fn eval_train(&self, params:&[Float], _:usize) -> Float
     {
         self.eval_test(params)
     }
